Add base proof verifier fail test

+++ src/main/java/com/apicatalog/ld/signature/ecdsa/sd/primitive/DerivedProofValue.java
@@ -60,7 +60,7 @@ public class DerivedProofValue implements ProofValue {
 
         final ByteArrayInputStream is = new ByteArrayInputStream(signature);
 
-        if (is.read() != BYTE_PREFIX[0] || is.read() != BYTE_PREFIX[1] || is.read() != BYTE_PREFIX[2]) {
+        if ((byte)is.read() != BYTE_PREFIX[0] || is.read() != BYTE_PREFIX[1] || is.read() != BYTE_PREFIX[2]) {
             throw new DocumentError(ErrorType.Invalid, "ProofValue");
         }
